+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="3.0.1"></a>
+## 3.0.1 (2016-08-06)
+
+
+### Bug Fixes
+
+* **validation:unique:** add support for whereNot ([54669d3](https://github.com/adonisjs/adonis-validation-provider/commit/54669d3))
+
+
+### Features
+
+* **validator:** add sanitizor and raw validations ([5efec8e](https://github.com/adonisjs/adonis-validation-provider/commit/5efec8e))
+* **validator:** add unique rule using database provider ([13a9b47](https://github.com/adonisjs/adonis-validation-provider/commit/13a9b47))
+
+
+
 <a name="3.0.0"></a>
 # 3.0.0 (2016-06-26)